reorganise component in folders
