+++ src/app/contact/contact-details/contact-details.component.scss
@@ -1,4 +1,4 @@
-@import "../../theme/variables";
+@import "../../../theme/variables";
 :host {
   display: flex;
   flex-direction: row;
