+++ src/app/contact/contact-dialog/contact-dialog.component.scss
@@ -1,4 +1,4 @@
-@import "../../theme/variables";
+@import "../../../theme/variables";
 :host {
   display: flex;
   flex-direction: column;
