+++ src/app/contact/contact.service.ts
@@ -20,19 +20,21 @@ export class ContactService {
   private _contacts$ = new BehaviorSubject<Contact[]>(server.data);
 
   getContacts(): Observable<Contact[]> {
-    return this._contacts$.asObservable().pipe(
-      map((list) =>
-        list.sort(function (a, b) {
-          if (a.surname < b.surname) {
-            return -1;
-          }
-          if (a.surname > b.surname) {
-            return 1;
-          }
-          return 0;
-        })
-      )
-    );
+    return this._contacts$
+      .asObservable()
+      .pipe(map((list) => this.sortContact(list)));
+  }
+
+  private sortContact(list: Contact[]) {
+    return list.sort(function (a, b) {
+      if (a.surname.toLowerCase() < b.surname.toLowerCase()) {
+        return -1;
+      }
+      if (a.surname.toLowerCase() > b.surname.toLowerCase()) {
+        return 1;
+      }
+      return 0;
+    });
   }
 
   setContacts(contacts: Contact[]) {
@@ -42,7 +44,8 @@ export class ContactService {
   addContact(contact: Contact) {
     const list = this._contacts$.value;
     list.push(contact);
-    this.setContacts(list);
+    const sortedList = this.sortContact(list);
+    this.setContacts(sortedList);
   }
 
   deleteContact(id: string) {
